Move Print tree func to a separate package
